refactor: available-classes-tabs 컴포넌트 추출
