+++ apps/service/src/pages/index.ts
@@ -2,7 +2,7 @@ export * from "./apply-mentor.page";
 export * from "./auth/login.page";
 export * from "./auth/signup-complete.page";
 export * from "./auth/signup.page";
-export * from "./available-classes.page";
+export * from "./available-classes/available-classes.page";
 export * from "./home/home.page";
 export * from "./mentor/dashboard.page";
 export * from "./my.page";
